+++ CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ### Changes
 
-- changed fights table to using react-bootstrap-table
+- moved fights table to it's own component
+- changed the style of events page (fights)
+- changed the file fighters to a bigger one
 
 ### Additions
 
--
+- added search to fights table